chore: Update intro/outro on response and move type
